+++ event-sourcing/event-publisher/published-event-processor/src/main/java/uk/gov/justice/services/eventsourcing/publishedevent/prepublish/EventPrePublisher.java
@@ -20,6 +20,8 @@ import javax.inject.Inject;
 import javax.sql.DataSource;
 import javax.transaction.Transactional;
 
+import org.slf4j.Logger;
+
 public class EventPrePublisher {
 
     @Inject