+++ event-sourcing/event-publisher/published-event-processor/src/test/java/uk/gov/justice/services/eventsourcing/publishedevent/rebuild/integration/RebuildPerformanceIT.java
@@ -24,6 +24,7 @@ import uk.gov.justice.services.eventsourcing.repository.jdbc.eventstream.EventSt
 import uk.gov.justice.services.jdbc.persistence.JdbcResultSetStreamer;
 import uk.gov.justice.services.jdbc.persistence.PreparedStatementWrapperFactory;
 import uk.gov.justice.services.messaging.DefaultJsonObjectEnvelopeConverter;
+import uk.gov.justice.services.messaging.spi.DefaultEnvelopeProvider;
 import uk.gov.justice.services.messaging.spi.DefaultJsonEnvelopeProvider;
 import uk.gov.justice.services.test.utils.persistence.DatabaseCleaner;
 import uk.gov.justice.services.test.utils.persistence.FrameworkTestDataSourceFactory;
@@ -99,7 +100,8 @@ public class RebuildPerformanceIT {
             StringToJsonObjectConverter.class,
             ObjectMapperProducer.class,
             LoggerProducer.class,
-            DefaultJsonEnvelopeProvider.class
+            DefaultJsonEnvelopeProvider.class,
+            DefaultEnvelopeProvider.class
     })
     @Default
     public WebApp war() {